# apps/api/openapi/api-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/orders": {
      "post": {
        "operationId": "BuyerOrderController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "BuyerOrder"
        ]
      },
      "get": {
        "operationId": "BuyerOrderController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING_PAYMENT",
                "PAID",
                "PARTIALLY_SHIPPED",
                "SHIPPED",
                "COMPLETED",
                "CANCELLED",
                "AFTER_SALE",
                "CLOSED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "BuyerOrder"
        ]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "BuyerOrderController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "BuyerOrder"
        ]
      }
    },
    "/api/v1/orders/{id}/tracking": {
      "get": {
        "operationId": "BuyerOrderController_tracking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "BuyerOrder"
        ]
      }
    },
    "/api/v1/orders/{id}/cancel": {
      "patch": {
        "operationId": "BuyerOrderController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "BuyerOrder"
        ]
      }
    },
    "/api/v1/orders/{id}/confirm-receipt": {
      "patch": {
        "operationId": "BuyerOrderController_confirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "BuyerOrder"
        ]
      }
    },
    "/api/v1/admin/orders": {
      "get": {
        "operationId": "AdminOrderController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING_PAYMENT",
                "PAID",
                "PARTIALLY_SHIPPED",
                "SHIPPED",
                "COMPLETED",
                "CANCELLED",
                "AFTER_SALE",
                "CLOSED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "AdminOrder"
        ]
      }
    },
    "/api/v1/admin/orders/{id}": {
      "get": {
        "operationId": "AdminOrderController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "AdminOrder"
        ]
      }
    },
    "/api/v1/admin/orders/{id}/tracking": {
      "get": {
        "operationId": "AdminOrderController_tracking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "AdminOrder"
        ]
      }
    },
    "/api/v1/admin/orders/{id}/merchant-orders/{merchantOrderId}/note": {
      "patch": {
        "operationId": "AdminOrderController_note",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "merchantOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantOrderNoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "AdminOrder"
        ]
      }
    },
    "/api/v1/admin/orders/{id}/merchant-orders/{merchantOrderId}/shipments": {
      "post": {
        "operationId": "AdminOrderController_ship",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "merchantOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShipmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "AdminOrder"
        ]
      }
    },
    "/api/v1/admin/orders/exports": {
      "post": {
        "operationId": "AdminOrderController_export",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderListQueryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "AdminOrder"
        ]
      }
    },
...
  },
  "info": {
    "title": "MoeCraft API",
    "description": "MoeCraft multi-merchant marketplace API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SubmitOrderDto": {
        "type": "object",
        "properties": {
          "quoteId": {
            "type": "string",
            "format": "uuid"
          },
          "signature": {
            "type": "string",
            "minLength": 20,
            "maxLength": 100
          }
        },
        "required": [
          "quoteId",
          "signature"
        ]
      },
      "MerchantOrderNoteDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "required": [
          "note"
        ]
      },
      "ShipmentLineDto": {
        "type": "object",
        "properties": {
          "orderItemId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 100000
          }
        },
        "required": [
          "orderItemId",
          "quantity"
        ]
      },
      "CreateShipmentDto": {
        "type": "object",
        "properties": {
          "carrier": {
            "type": "string",
            "enum": [
              "SF",
              "ZTO",
              "YTO",
              "STO",
              "YD",
              "JD",
              "EMS"
            ]
          },
          "trackingNumber": {
            "type": "string",
            "pattern": "^[A-Za-z0-9-]{4,80}$"
          },
          "note": {
            "type": "string",
            "maxLength": 500
          },
          "items": {
            "minItems": 1,
            "maxItems": 200,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShipmentLineDto"
            }
          }
        },
        "required": [
          "carrier",
          "trackingNumber",
          "items"
        ]
      },
      "OrderListQueryDto": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "PENDING_PAYMENT",
              "PAID",
              "PARTIALLY_SHIPPED",
              "SHIPPED",
              "COMPLETED",
              "CANCELLED",
              "AFTER_SALE",
              "CLOSED"
            ],
            "type": "string"
          },
          "search": {
            "type": "string",
            "maxLength": 80
          }
        }
      },
...
    }
  }
}